Gestión de fases, pero al modificar fase de hito ERROR

+++ resources/js/components/FaseComponent.vue
@@ -18,26 +18,59 @@
     </form>
 
     <table class="table table-hover">
-        <thead>
+        <thead class="thead-light">
             <tr>
-                <th class="text-center" scope="col">ID</th>
-                <th class="text-center" scope="col">NOMBRE</th>
-                <th class="text-center" scope="col">DESCRIPCION</th>
-                <th class="text-center" scope="col">OPCIONES</th>
+                <th class="text-center" scope="col" width="5%">ID</th>
+                <th class="text-center" scope="col" width="20%">NOMBRE</th>
+                <th class="text-center" scope="col" width="30%">DESCRIPCION</th>
+                <th class="text-center" scope="col" width="35%">OPCIONES</th>
             </tr>
         </thead>
         <tbody>
             <tr v-for="(item, index) in fases" v-bind:key="index">
                 <td class="text-center">{{item.id}}</td>
-                <td class="text-center">{{item.nombre}}</td>
-                <td class="text-center">{{item.descripcion}}</td>
+                <td>{{item.nombre}}</td>
+                <td>{{item.descripcion}}</td>
                 <td class="text-center">
                     <button type="button" @click="editarFormulario(item)" class="btn btn-primary">Modificar</button>
+                    <button type="button" @click="relacionHitos(item)" class="btn btn-info">Relación con hitos</button>
                     <button type="button" @click="eliminarFase(item, index)" class="btn btn-secondary">Eliminar</button>
                 </td>
             </tr>
         </tbody>
     </table>
+
+    <b-modal size="xl" id="modal-hitos" ref="btnUser" hide-footer no-close-on-esc hide-header-close>
+        <template v-slot:modal-title>
+            Fase a eliminar: <b> {{fase.id}} - {{fase.nombre}} </b>
+        </template>
+        <div>
+            <p>No se puede eliminar, está relacionada con los siguientes atributos: </p>
+        </div>
+        <table class="table table-hover">
+            <thead class="thead-light">
+                <tr>
+                    <th class="text-center" scope="col">HITO</th>
+                    <th class="text-center" scope="col">ELEGIR OTRO HITO</th>
+                </tr>
+            </thead>
+            <tbody>
+                <tr v-for="(hito_coincidente, index) in hito_coincidentes" v-bind:key="index">
+                    <td>{{hito_coincidente.id}} - {{hito_coincidente.nombre}}</td>
+                    <td class="text-center">
+                        <b-form-select v-model="hito_coincidente.fase_id" v-on:change="guardarCambios(hito_coincidente, index)">
+                            <option v-for="hito in hitos" v-bind:key="hito.id" :value="hito.id">{{hito.id}} - {{hito.nombre}}</option>
+                        </b-form-select>
+                    </td>
+                </tr>
+            </tbody>
+        </table>
+        <b-row>
+            <b-col>
+                <b-button class="mt-3" block @click="cancelarEdicionModal()"> Cerrar </b-button>
+            </b-col>
+        </b-row>        
+    </b-modal>
 
 </div> 
 </template>
@@ -54,16 +87,35 @@ export default {
                 nombre: '',
                 descripcion: ''
             },
-            editarActivo: false
+            editarActivo: false,
+            hitos: [],
+            hito: 
+            {
+                id: '',
+                fase_id: '',
+                nombre: '',
+                descripcion: ''
+            },
+            hito_coincidentes: [],
+            hito_coincidente: 
+            {
+                id: '',
+                fase_id: '',
+                nombre: '',
+                descripcion: ''
+            }
         }
     },
     created()
     {
-        console.log('FaseComponent created')
         axios.get('/admin/fases/create')
             .then(res=>{
                 this.fases = res.data;
             });
+        axios.get('/admin/hitos/create')
+            .then(res=>{
+                this.hitos = res.data;
+            })
     },
     methods:
     {
@@ -96,11 +148,43 @@ export default {
                     this.fase = {nombre:'', descripcion:''};
                 })
         },
+        relacionHitos(item)
+        {
+            this.fase.nombre = item.nombre;
+            this.fase.descripcion = item.descripcion;
+            this.fase.id = item.id;
+
+            for(var i=0; i<this.hitos.length; i++)
+            {
+                if(this.hitos[i].fase_id === item.id)
+                {
+                    this.hito_coincidentes.push(this.hitos[i]);
+                }
+            }
+            this.$root.$emit('bv::show::modal', 'modal-hitos', '#btnShow')
+        },
         eliminarFase(fase, index)
         {
-            axios.delete(`/admin/fases/${fase.id}`)
-                .then(()=>{
-                    this.fases.splice(index, 1);
+            axios.get(`/admin/hitos/${fase.id}/edit`)
+                .then(res=>{
+                    if(res.data === 0)
+                    {
+                        this.$bvModal.msgBoxConfirm("¿Quiere eliminar?",{
+                            okVariant: 'danger',
+                            okTitle: 'Eliminar',
+                            cancelTitle: 'Cancelar'
+                        }).then(value=>{
+                            if(value === true)
+                            {
+                                axios.delete(`/admin/fases/${fase.id}`)
+                                    .then(()=>{
+                                        this.fases.splice(index, 1);
+                                    })
+                            }
+                        })
+                    }else{
+                        alert("La fase tiene hitos relacionados, no se puede eliminar")
+                    }
                 })
         },
         cancelarEdicion()
@@ -108,12 +192,30 @@ export default {
             this.editarActivo = false;
             this.fase = {nombre: '', descripcion: ''};
         },
+        cancelarEdicionModal()
+        {
+            this.fase = {nombre: '', descripcion: ''};
+            this.hito_coincidentes = [];
+            this.$root.$emit('bv::hide::modal', 'modal-hitos')
+        },
         editarFormulario(item)
         {
             this.editarActivo = true;
             this.fase.nombre = item.nombre;
             this.fase.descripcion = item.descripcion;
             this.fase.id = item.id;
+        },
+        guardarCambios(item, index)
+        {
+            const params = {fase_id: item.fase_id, nombre: item.nombre, descripcion: item.descripcion}
+            axios.put(`/admin/hitos/${item.id}`, params)
+                .then(res=>{
+                    const index = this.hitos.findIndex(hitoBuscar => hitoBuscar.id === res.data.id)
+                    this.hitos[index] = res.data;
+                });
+            this.$root.$emit('bv::hide::modal', 'modal-hitos', '#btnShow');
+            this.fase = {nombre: '', descripcion: ''};
+            this.hito_coincidentes = [];
         }
     }
 }
